refactor: products-write hook 파일들을 hooks 폴더로 이동 및 구조화
